+++ doc/CHANGELOG.md
@@ -26,6 +26,7 @@ For release instructions, see [cynic-net/pypi-release] on GitHub.
 
 
 ### dev
+- Added: Simulator/unit test BIOS source code and `t8dev aslt8dev` command
 
 ### 0.0.4 (2024-09-22)
 - Fixed: `t8dev emu` no longer tries to use `wine` on Windows